Complete 2nd Foundation Lesson
I learned how to apply ID and Class selectors. I also missed that comma
was supposed to be a semi-colon for over 30 minutes. Which in my opinion
was ridiculous. That was my first typo problem in coding so far. I
laughed, but it will not be funny AT ALL in future tense XD! I als got
to work with font families.